Fixed #3122:
Now the forms framework labels, error messages and help texts can be easily translated setting a php callable (or a sfCallable instance) for this purpose. In a symfony context (remember that the new forms framework can be used without symfony), the callable (`__()`) is automatically set through the default factories configuration process.

 * Added a constructor to the `sfWidgetFormSchemaFormatter`, which from now must must have an instance of `sfWidgetFormSchema` instance passed as a parameter,
 * Added a way to statically get and set a translation callable in `sfWidgetFormSchemaFormatter`,
 * Moved html label generation methods from `sfWidgetFormSchema` to `sfWidgetFormSchemaFormatter` (but this won't break BC),
 * Added translation callable declaration in default `factories.yml` symfony i18n configuration,
 * Created according unit and functional tests.
 

git-svn-id: http://svn.symfony-project.com/branches/1.1@8408 ee427ae8-e902-0410-961c-c3ed070cd9f9